Add relations between users and tasks

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { GetTasksFilter } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './entities/task.entity';
+import { User } from './../auth/entities/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -15,32 +16,37 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) { }
 
-    async getAllTasks(filterDto: GetTasksFilter): Promise<Array<Task>> {
-        return await this.taskRepository.getTasks(filterDto);
+    async getAllTasks(filterDto: GetTasksFilter, user: User): Promise<Array<Task>> {
+        return await this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
         if (!found) {
             throw new NotFoundException(`Task with ID = ${id} not found!`);
         }
         return found
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number): Promise<{result: boolean}> {
-        const queryResult = await this.taskRepository.delete(id);
-        return { result: !!queryResult.affected };
+    async deleteTaskById(id: number, user: User): Promise<{result: boolean}> {
+        const queryResult = await this.taskRepository.delete({id, userId: user.id});
+        if(queryResult.affected < 1) {
+            throw new NotFoundException(`Task with ID = ${id} not found!`);
+        }
+        return { result: true };
     }
 
-    async updateTaskStatus(updateStatus: UpdateStatus): Promise<{result: boolean}> {
+    async updateTaskStatus(updateStatus: UpdateStatus, user: User): Promise<{result: boolean}> {
         const {id, status} = updateStatus;
-        const queryResult = await this.taskRepository.update(id, {status});
-        return { result: !!queryResult.affected };
-         
+        const task = await this.getTaskById(id, user);
+        if(!task) return { result: false }
+        task.status = status;
+        task.save();
+        return { result: true }
     }
 
 }
